+++ instana/resource-group_test.go
@@ -92,7 +92,7 @@ func TestCRUDOfMinimalRBACGroupResourceWithMockServer(t *testing.T) {
 		}
 		`
 
-	testStepsFactory := func(httpServerPort int, resourceID string) []resource.TestStep {
+	testStepsFactory := func(httpServerPort int64, resourceID string) []resource.TestStep {
 		return []resource.TestStep{
 			createMinimalRbacGroupResourceTestStep(httpServerPort, 0, resourceID),
 			testStepImportWithCustomID(testGroupDefinition, resourceID),
@@ -104,7 +104,7 @@ func TestCRUDOfMinimalRBACGroupResourceWithMockServer(t *testing.T) {
 	executeRBACGroupIntegrationTest(t, serverResponseTemplate, testStepsFactory)
 }
 
-func createMinimalRbacGroupResourceTestStep(httpPort int, iteration int, id string) resource.TestStep {
+func createMinimalRbacGroupResourceTestStep(httpPort int64, iteration int, id string) resource.TestStep {
 	return resource.TestStep{
 		Config: appendProviderConfig(fmt.Sprintf(minimalRBACGroupDefinitionTemplate, iteration), httpPort),
 		Check: resource.ComposeTestCheckFunc(
@@ -128,7 +128,7 @@ func TestCRUDOfRBACGroupResourceWithPermissionsAndApplicationIdsAssignedUsingMoc
 		}
 		`
 
-	testStepsFactory := func(httpServerPort int, resourceID string) []resource.TestStep {
+	testStepsFactory := func(httpServerPort int64, resourceID string) []resource.TestStep {
 		return []resource.TestStep{
 			createRbacGroupResourceWithPermissionsAndApplicationIdsAssignedTestStep(httpServerPort, 0, resourceID),
 			testStepImportWithCustomID(testGroupDefinition, resourceID),
@@ -140,7 +140,7 @@ func TestCRUDOfRBACGroupResourceWithPermissionsAndApplicationIdsAssignedUsingMoc
 	executeRBACGroupIntegrationTest(t, serverResponseTemplate, testStepsFactory)
 }
 
-func createRbacGroupResourceWithPermissionsAndApplicationIdsAssignedTestStep(httpPort int, iteration int, id string) resource.TestStep {
+func createRbacGroupResourceWithPermissionsAndApplicationIdsAssignedTestStep(httpPort int64, iteration int, id string) resource.TestStep {
 	return resource.TestStep{
 		Config: appendProviderConfig(fmt.Sprintf(rbacGroupDefinitionWithPermissionsAndApplicationIDsAssignedTemplate, iteration), httpPort),
 		Check: resource.ComposeTestCheckFunc(
@@ -176,7 +176,7 @@ func TestCRUDOfFullRBACGroupResourceWithMockServer(t *testing.T) {
 		}
 		`
 
-	testStepsFactory := func(httpServerPort int, resourceID string) []resource.TestStep {
+	testStepsFactory := func(httpServerPort int64, resourceID string) []resource.TestStep {
 		return []resource.TestStep{
 			createFullRbacGroupResourceTestStep(httpServerPort, 0, resourceID),
 			testStepImportWithCustomID(testGroupDefinition, resourceID),
@@ -188,7 +188,7 @@ func TestCRUDOfFullRBACGroupResourceWithMockServer(t *testing.T) {
 	executeRBACGroupIntegrationTest(t, serverResponseTemplate, testStepsFactory)
 }
 
-func createFullRbacGroupResourceTestStep(httpPort int, iteration int, id string) resource.TestStep {
+func createFullRbacGroupResourceTestStep(httpPort int64, iteration int, id string) resource.TestStep {
 	return resource.TestStep{
 		Config: appendProviderConfig(fmt.Sprintf(fullRBACGroupDefinitionTemplate, iteration), httpPort),
 		Check: resource.ComposeTestCheckFunc(
@@ -220,7 +220,7 @@ func testCheckPermissionSetStringSetField(attribute string, idx int, value strin
 	return resource.TestCheckResourceAttr(testGroupDefinition, fmt.Sprintf("%s.0.%s.%d", GroupFieldPermissionSet, attribute, idx), value)
 }
 
-type rbacGroupTestStepsFactory func(httpServerPort int, resourceID string) []resource.TestStep
+type rbacGroupTestStepsFactory func(httpServerPort int64, resourceID string) []resource.TestStep
 
 func executeRBACGroupIntegrationTest(t *testing.T, serverResponseTemplate string, testStepsFactory rbacGroupTestStepsFactory) {
 	id := RandomID()
